chore: add index.html .gitignore (base structure)
The `index.html file contains the basic structure of the website.

The `.gitignore` file contains the files and folders that I don't want
to upload to GitHub. In this case, I'm using a Python HTTP server to
view changes to the local website in real time in the browser, and I've
configured the logs from that Python HTTP server to be saved in a
nohup.out file, which I don't want uploaded to GitHub.